pyproject.toml
[tool.poetry]
name = "derm_tiff"
version = "2.0.7"
description = "Utilities for tiff images created by DermAnnotation"
authors = ["kawa-yo <yoshito.kawasaki.pub@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8, <4.0"
nptyping = "^2.2.0"
numpy = "^1.24.2"
opencv-python = "^4.7.0"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
